+++ R/set_fallback.R
@@ -1,30 +1,41 @@
-#' @title Set a Fallback Learner
+#' @title Create a Fallback Learner
 #'
 #' @description
-#' Set a fallback learner for a given learner.
+#' Create a fallback learner for a given learner.
 #' The function searches for a suitable fallback learner based on the task type.
 #' Additional checks are performed to ensure that the fallback learner supports the predict type.
 #'
 #' @param learner [Learner]\cr
-#'  The learner for which a fallback learner should be set.
+#'  The learner for which a fallback learner should be created.
 #'
-#' @return
-#' Returns the learner itself, but modified **by reference**.
-set_fallback = function(learner) {
-  assert_learner(learner)
+#' @return [Learner]
+default_fallback = function(learner, ...) {
+  UseMethod("default_fallback")
+}
 
-  # search for suitable fallback learner
-  fallback_id = mlr_reflections$learner_fallback[[learner$task_type]]
-
-  if (is.null(fallback_id)) {
-    stopf("No fallback learner available for task type '%s'.", learner$task_type)
-  }
-
-  fallback = lrn(fallback_id)
+#' @rdname default_fallback
+#' @export
+default_fallback.LearnerClassif = function(learner) {
+  fallback = lrn("classif.featureless")
 
   # set predict type
   if (learner$predict_type %nin% fallback$predict_types) {
     stopf("Fallback learner '%s' does not support predict type '%s'.", fallback_id, learner$predict_type)
+  }
+
+  fallback$predict_type = learner$predict_type
+
+  return(fallback)
+}
+
+#' @rdname default_fallback
+#' @export
+default_fallback.LearnerRegr = function(learner) {
+  fallback = lrn("regr.featureless")
+
+  # set predict type
+  if (learner$predict_type %nin% fallback$predict_types) {
+    stopf("Fallback learner '%s' does not support predict type '%s'.", fallback$id, learner$predict_type)
   }
 
   fallback$predict_type = learner$predict_type
@@ -40,6 +51,5 @@ set_fallback = function(learner) {
     fallback$quantile_response = learner$quantile_response
   }
 
-  learner$fallback = fallback
-  return(learner)
+  return(fallback)
 }